+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/RequestDto/lotteryEventDto/LotteryEventRequestDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.RequestDto;
+package JGS.CasperEvent.domain.event.dto.RequestDto.lotteryEventDto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
